Fix build
explicitly cast to int32_t to have correct std::pair comparison

+++ src/ku/constraintTemplated/utilities/NeighborPatternUtilities.hpp
@@ -297,7 +297,8 @@ computePatternOptionNonTorus(uint32_t rowIndex,
           if (otherNeighborOptional.has_value()) {
             backwardId++;
             const auto& otherNeighbor = otherNeighborOptional.value();
-            if (otherNeighbor == std::make_pair(rowIndex, columnIndex)) {
+            if (otherNeighbor ==
+                std::make_pair(static_cast<std::int32_t>(rowIndex), static_cast<std::int32_t>(columnIndex))) {
               break;
             }
           }
